Pure library (#13)
* Pure CSS

Added a CSS library with grids
Updated the Work section to incorporate new CSS grib
Added links and set up 6 work projects

* Saved Style Changes

Last comitt did not include CSS file changes.

* Fixed CSS issue

There was an issue with CSS that was causing problems this weekend. The image class was not repsonsive. I fixed that issue.