Link Styling, Header Padding, Welcome Positioning
I added a link to work overview page and updated the styling on the link. Set hover state to show dropshadow, might want to change that.
Decreased the top and bottom padding on the header items.
I adjusted the alignment of the welcome message so that it's positioned closer to the top of the screen on desktop and mobile.